remove `permute` and disable padding in add deltas layer.
